fix: Refactored table styling
