+++ src/components/workouts/workoutExercise/WorkoutList.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, View } from "react-native";
 import React, { FC } from "react";
-import { WorkoutExercises } from "../../types/workouts";
+import { WorkoutExercises } from "../../../types/workouts";
 import { useTheme } from "react-native-paper";
 import WorkoutExerciseItem from "./WorkoutExerciseItem";
